Ensure target is an instance of the test class for which this FieldTypeRedefinitions was created
Discovered when upgrading to junit 6 that invalid data can come through which then results in a lot of duplicated errors with stacks that are not worthwhile.  The same data is produced by better capture and affects far less tests as a result.  In this case, 10 failed prior but only 2 were in error when handled better here.  This doesn't fix issues with junit 6, but helps in limited leaks.