Bump agent to 2.0.0 and add seeding logic to ensure m2 cache has it so all arglines work
On github actions, I had been working around this problem by performing install.  This trick effectively does the same allowing me to remove the 'install' phase entirely there.  Another option would have been to add the dependency into coverage tests and elsewhere but this ensures its simply there during the build effectively solving for all.  I had a note as well to use gmavenplus which would be fine if I didn't know if downstream needed or not but in this case i know so this works even better.